virtual_tryon/requirements.txt
torch
torchvision
pillow
numpy
opencv-python
flask
